Minimal work to auth against Azure AD in Angular
Here, we add the angular-oauth2-oidc module, import its service,
and configure it for our tenant and app. We also add a route with
a RouteGuard that ensures the user authenticates before going to
that route.

We've also added an API call where we'll use our token, but it
doesn't work yet. Authentication itself works, but we have no
back end yet.