Add back end API requiring JWT
With this commit, we add the API call on the back end and make changes
to the Startup class to support web API, authentication, and serving
static files, while routing anything else to our front end. We also
update the front end to pass the token when we make the API call.

One other change to note is the proxy config. The way we're set up now,
you should start the back end by running `dotnet run` from the folder
with the csproj in it. This launches the web API listening on port
5000. Then, launch the front end by running `npm start` from the
ClientApp folder. The front end will listen on 4200, and will proxy
any api calls to port 5000.

In other words, you launch the front end and back end separately
and debug them separately.